liuxu

+++ src/edu/hbuas/LandDiscover/model/javaBean/Account.java
@@ -3,7 +3,7 @@ package edu.hbuas.LandDiscover.model.javaBean;
 
 public class Account {
 
-  private long userId;
+  private int userId;
   private String username;
   private String password;
   private String nickname;
@@ -14,7 +14,7 @@ public class Account {
   public Account() {
   }
 
-  public Account(long userId, String username, String password, String nickname, long phone, String sex, String address) {
+  public Account(int userId, String username, String password, String nickname, long phone, String sex, String address) {
     this.userId = userId;
     this.username = username;
     this.password = password;
@@ -24,11 +24,11 @@ public class Account {
     this.address = address;
   }
 
-  public long getUserId() {
+  public int getUserId() {
     return userId;
   }
 
-  public void setUserId(long userId) {
+  public void setUserId(int userId) {
     this.userId = userId;
   }
